Refactor and change API to simplify
